+++ esphome/components/ota_http/ota_http_idf.h
@@ -12,7 +12,7 @@ namespace ota_http {
 class OtaHttpIDF : public OtaHttpComponent {
  public:
   int http_init() override;
-  size_t http_read(uint8_t *buf, size_t len) override;
+  int http_read(uint8_t *buf, size_t len) override;
   void http_end() override;
 
  protected:
